m3/requirements.txt
llm
python-dotenv
llm-claude-3
pytest